# docs/05-api/api-spec.yaml
openapi: 3.0.3
info:
  title: EduCenter LMS API
  version: 1.0.0
  description: |
    API chuẩn cho EduCenter LMS (on-premise). Mọi endpoint yêu cầu JWT bearer
    trừ nhóm /public, /auth/login, /auth/forgot-password, /auth/reset-password,
    /health và webhook tích hợp.

    Quy ước chung:
    - Base URL: https://<host>/api (reverse proxy qua Nginx)
    - Timestamp: ISO 8601 (UTC)
    - Tiền tệ: NUMERIC decimal, đơn vị theo organization.currency (mặc định VND)
    - Lỗi: { "code", "message", "correlation_id", "details" }
    - Pagination: ?page=1&page_size=50 → meta.total, meta.page, meta.page_size
    - Idempotency: endpoint ghi dữ liệu tài chính/ghi danh nhận header Idempotency-Key
    - Rate limit: 100 req/15 phút/IP; login 5 lần/15 phút
    - Mọi endpoint business check scope (branch/class/student) + module effective state

    Endpoint đánh dấu x-phase: thuộc roadmap (P2/P3), không nằm trong phạm vi MVP.
  contact:
    name: EduCenter LMS Vendor
    email: support@lms-vendor.com
  license:
    name: Proprietary
servers:
  - url: https://lms.example.edu.vn/api
    description: Production (customer installation)
...
paths:
...
  /departments:
    get:
      tags: [Academic]
      summary: Danh sách ngành
      operationId: listDepartments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách ngành
    post:
      tags: [Academic]
      summary: Tạo ngành
      operationId: createDepartment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name]
              properties:
                code:
                  type: string
                name:
                  type: string
      responses:
        "201":
          description: Đã tạo
  /programs:
    get:
      tags: [Academic]
      summary: Danh sách chương trình
      operationId: listPrograms
      security:
        - bearerAuth: []
      parameters:
        - name: department_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
      responses:
        "200":
          description: Danh sách chương trình
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Program"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Academic]
      summary: Tạo chương trình
      operationId: createProgram
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [department_id, code, name]
              properties:
                department_id:
                  type: string
                  format: uuid
                code:
                  type: string
                name:
                  type: string
                description:
                  type: string
                duration_months:
                  type: integer
                completion_rules:
                  type: object
      responses:
        "201":
          description: Đã tạo

  # DEVIATION (T040): /courses + /rooms + PUT/GET chi tiết + DELETE schedule thêm ngoài bản
  # api-spec gốc — theo DDL docs/04-database-schema.md §6 (courses thuộc chương trình, rooms
  # phục vụ lịch học). Cấu trúc khóa học không có màn hình riêng trong mockup (tab Khóa học).
  /courses:
    get:
      tags: [Academic]
      summary: Danh sách khóa học (lọc program_id/status)
      operationId: listCourses
      security:
        - bearerAuth: []
      parameters:
        - name: program_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
      responses:
        "200":
          description: Danh sách khóa học
    post:
      tags: [Academic]
      summary: Tạo khóa học
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [program_id, code, name]
              properties:
                program_id:
                  type: string
                  format: uuid
                code:
                  type: string
                name:
                  type: string
                description:
                  type: string
                order_index:
                  type: integer
      responses:
        "201":
          description: Đã tạo
    put:
      tags: [Academic]
      summary: Cập nhật khóa học
      operationId: updateCourse
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                order_index:
                  type: integer
                status:
                  type: string
                  enum: [draft, active, archived]
      responses:
        "200":
          description: Đã cập nhật
  /rooms:
    get:
      tags: [Academic]
      summary: Danh sách phòng học (branch scope)
      operationId: listRooms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Danh sách phòng
    post:
      tags: [Academic]
      summary: Tạo phòng học
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_id, code]
              properties:
                branch_id:
                  type: string
                  format: uuid
                code:
                  type: string
                name:
                  type: string
                capacity:
                  type: integer
      responses:
        "201":
          description: Đã tạo
  /classes/{class_id}:
    get:
      tags: [Academic]
      summary: Chi tiết lớp (kèm giảng viên)
      operationId: getClass
      security:
        - bearerAuth: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chi tiết lớp
    put:
      tags: [Academic]
      summary: Cập nhật lớp
      operationId: updateClass
      security:
        - bearerAuth: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                modality:
                  type: string
                  enum: [offline, online, hybrid, flexible]
                capacity:
                  type: integer
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                enrollment_status:
                  type: string
                  enum: [draft, open, closed, full, archived]
                status:
                  type: string
                  enum: [draft, active, archived]
                teacher_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Đã cập nhật
  /classes:
    get:
      tags: [Academic]
      summary: Danh sách lớp học
      operationId: listClasses
      security:
        - bearerAuth: []
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: program_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, open, closed, full, archived]
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Danh sách lớp
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Class"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Academic]
      summary: Tạo lớp học
      operationId: createClass
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_id, program_id, course_id, code, name]
              properties:
                branch_id:
                  type: string
                  format: uuid
                program_id:
                  type: string
                  format: uuid
                course_id:
                  type: string
                  format: uuid
                code:
                  type: string
                name:
                  type: string
                modality:
                  type: string
                  enum: [offline, online, hybrid, flexible]
                capacity:
                  type: integer
                  minimum: 1
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                teacher_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "409":
          description: Mã lớp trùng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /classes/{class_id}/schedules:
    parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Academic]
      summary: Tạo lịch học (chống trùng teacher/room/time)
      operationId: createSchedule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [day_of_week, start_time, end_time, teacher_id, valid_from]
              properties:
                day_of_week:
                  type: integer
                  minimum: 1
                  maximum: 7
                start_time:
                  type: string
                  example: "18:00"
                end_time:
                  type: string
                  example: "19:30"
                teacher_id:
                  type: string
                  format: uuid
                room_id:
                  type: string
                  format: uuid
                recurrence:
                  type: string
                  enum: [weekly, biweekly, once]
                valid_from:
                  type: string
                  format: date
                valid_to:
                  type: string
                  format: date
      responses:
        "201":
          description: Đã tạo lịch
        "409":
          description: Trùng lịch (teacher/room/time)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Academic]
      summary: Xóa lịch học
      operationId: deleteSchedule
      security:
        - bearerAuth: []
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: schedule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Đã xóa

  # ============================================================
  # STUDENTS & ENROLLMENT
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        correlation_id:
          type: string
          format: uuid
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar_url:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, deleted]
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        created_at:
          type: string
          format: date-time
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        is_system:
          type: boolean
        permissions:
          type: array
          items:
            type: string
...
    Program:
      type: object
      properties:
        id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        duration_months:
          type: integer
        completion_rules:
          type: object
        status:
          type: string
          enum: [draft, active, archived]
    Class:
      type: object
      properties:
        id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
        program_id:
          type: string
          format: uuid
        course_id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        modality:
          type: string
          enum: [offline, online, hybrid, flexible]
        capacity:
          type: integer
        enrolled_count:
          type: integer
        enrollment_status:
          type: string
          enum: [draft, open, closed, full, archived]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        teachers:
          type: array
          items:
            $ref: "#/components/schemas/User"
...